small visual fixes

+++ app/Http/Controllers/ChapterController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Chapter;
 use App\Models\Course;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use function PHPUnit\Framework\returnArgument;
 
 class ChapterController extends Controller
@@ -38,26 +39,30 @@ class ChapterController extends Controller
      */
     public function store(Request $request, Course $course)
     {
-        $request->validate([
-            'name'=>'required|max:15|unique:chapters,name,NULL,id,course_id,' . $course->id,
-            'pages'=>'required|numeric|min:0'
-        ]);
-        $chapter = new Chapter();
-        $chapter->setStatus('not-started');
-        $chapter->setPages(request('pages'));
-        $chapter->setName(request('name'));
-        $chapter->setCourse($course->id);
-        $chapter->save();
-        if(request('action_url')){
-            return redirect(request('action_url'))->with('message', 'New chapter ' . $chapter->getName() . ' added.');
+        if ($course->period->user_id == Auth::id()) {
+            $request->validate([
+                'name' => 'required|max:15|unique:chapters,name,NULL,id,course_id,' . $course->id,
+                'pages' => 'required|numeric|min:0'
+            ]);
+            $chapter = new Chapter();
+            $chapter->setStatus('not-started');
+            $chapter->setPages(request('pages'));
+            $chapter->setName(request('name'));
+            $chapter->setCourse($course->id);
+            $chapter->save();
+            if (request('action_url')) {
+                return redirect(request('action_url'))->with('message', 'New chapter ' . $chapter->getName() . ' added.');
+            }
+            return redirect('/periods/' . $course->period->id)->with('message', 'New chapter added for: ' . $course->getName());
         }
-        return redirect('/periods/' . $course->period->id)->with('message', 'New chapter added for: ' . $course->getName());
+        abort(403);
     }
+
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Chapter  $chapter
+     * @param \App\Models\Chapter $chapter
      * @return \Illuminate\Http\Response
      */
     public function show(Chapter $chapter)
@@ -68,7 +73,7 @@ class ChapterController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Chapter  $chapter
+     * @param \App\Models\Chapter $chapter
      * @return \Illuminate\Http\Response
      */
     public function edit(Chapter $chapter)
@@ -79,60 +84,64 @@ class ChapterController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Chapter  $chapter
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Models\Chapter $chapter
      *
      */
     public function update(Request $request, Chapter $chapter)
     {
-        if(request('name')=== $chapter->name){
-            $request->validate([
-                'name'=>'required|max:15',
-                'pages'=>'required|numeric|min:0'
-            ]);
-        } else{
-            $request->validate([
-                'name'=>'required|max:15|unique:chapters,name,NULL,id,course_id,' . $chapter->course->id,
-                'pages'=>'required|numeric|min:0'
-            ]);
+        if ($chapter->course->period->user_id == Auth::id()) {
+            if (request('name') === $chapter->name) {
+                $request->validate([
+                    'name' => 'required|max:15',
+                    'pages' => 'required|numeric|min:0'
+                ]);
+            } else {
+                $request->validate([
+                    'name' => 'required|max:15|unique:chapters,name,NULL,id,course_id,' . $chapter->course->id,
+                    'pages' => 'required|numeric|min:0'
+                ]);
+            }
+            $chapter->setPages(request('pages'));
+            $chapter->setName(request('name'));
+            $chapter->save();
+            return redirect('/courses/show/' . $chapter->course->id)->with('message', 'Chapter ' . $chapter->getName() . ' was updated.');
         }
-        $request->validate([
-            'name'=>'required|max:15|unique:chapters,name,NULL,id,course_id,' . $chapter->course->id,
-            'pages'=>'required|numeric|min:0'
-        ]);
-        $chapter->setPages(request('pages'));
-        $chapter->setName(request('name'));
-        $chapter->save();
-        return redirect('/courses/show/' . $chapter->course->id)->with('message', 'Chapter ' . $chapter->getName() . ' was updated.');
+        abort(403);
     }
+
 
     public function updateStatus(Request $request, Chapter $chapter)
     {
-        $request->validate([
-            'status' => 'required|in:not-started,busy,done',
-        ]);
-        if(request('action_url')){
-
+        if ($chapter->course->period->user_id == Auth::id()) {
+            $request->validate([
+                'status' => 'required|in:not-started,busy,done',
+            ]);
+            $chapter->setStatus(request('status'));
+            $chapter->save();
+            $course = $chapter->course;
+            if (request('action_url')) {
+                return redirect(request('action_url'))->with('message', 'Status for ' . $chapter->getName() . ' set to ' . $chapter->getStatus());
+            }
+            return redirect('/periods/' . $course->period->id)->with('message', 'Status for ' . $chapter->getName() . ' set to ' . $chapter->getStatus());
         }
-        $chapter->setStatus(request('status'));
-        $chapter->save();
-        $course = $chapter->course;
-        if(request('action_url')){
-            return redirect(request('action_url'))->with('message', 'Status for ' . $chapter->getName() . ' set to ' . $chapter->getStatus());
-        }
-        return redirect('/periods/' . $course->period->id)->with('message', 'Status for ' . $chapter->getName() . ' set to ' . $chapter->getStatus());
+        abort(403);
     }
+
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Chapter  $chapter
+     * @param \App\Models\Chapter $chapter
      * @return \Illuminate\Http\Response
      */
     public function destroy(Chapter $chapter)
     {
-        $courseID = $chapter->course->id;
-        $name = $chapter->getName();
-        $chapter->delete();
-        return redirect('/courses/show/' . $courseID)->with('message', $name . ' was deleted.');
+        if ( $chapter->course->period->id == Auth::id() ) {
+            $courseID = $chapter->course->id;
+            $name = $chapter->getName();
+            $chapter->delete();
+            return redirect('/courses/show/' . $courseID)->with('message', $name . ' was deleted.');
+        }
+
     }
 }
